coagents-pinn-solver/agent/pyproject.toml
[tool.poetry]
name = "pinn_solver"
version = "0.1.0"
description = "Physics-Informed Neural Network Solver with CopilotKit"
authors = ["OpenHands AI <contact@openhands.ai>"]
license = "MIT"

[project]
name = "pinn_solver"
version = "0.0.1"
dependencies = [
    "copilotkit>=0.1.52",
    "langchain-openai>=0.2.3",
    "langchain-community>=0.3.1",
    "langchain>=0.3.4",
    "openai>=1.52.1",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.31.0",
    "requests>=2.32.3",
    "langchain-core>=0.3.25",
    "langgraph-cli[inmem]>=0.3.3",
    "langgraph-checkpoint-sqlite>=2.0.1",
    "aiosqlite>=0.20.0",
    "aiohttp>=3.9.3",
    "langgraph==0.4.8",
    "boto3>=1.34.0",
    "numpy>=1.24.3",
    "matplotlib>=3.8.2",
    "pandas>=2.1.4",
    "fastapi>=0.104.1",
    "pydantic>=2.5.0"
]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
copilotkit = "^0.1.52"
langchain-openai = "0.2.3"
langchain-community = "^0.3.1"
langchain = "0.3.26"
openai = "^1.52.1"
python-dotenv = "^1.0.1"
uvicorn = "^0.31.0"
requests = "^2.32.3"
langchain-core = "^0.3.25"
langgraph-cli = {extras = ["inmem"], version = "^0.3.3"}
langgraph-checkpoint-sqlite = "^2.0.1"
aiosqlite = "^0.20.0"
aiohttp = "^3.9.3"
langgraph = "0.4.8"
boto3 = "^1.34.0"
numpy = "^1.24.3"
matplotlib = "^3.8.2"
pandas = "^2.1.4"
fastapi = "^0.104.1"
pydantic = "^2.5.0"

[tool.poetry.scripts]
demo = "pinn_solver.demo:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
flake8 = "^6.1.0"